requirements.txt
google-adk
numexpr
scipy
pint
periodictable
duckduckgo_search
